refactor(upload): normalize code style and reorganize files
